+++ Sources/CBzip2/shim.h
@@ -1,1 +1,2 @@
 #include <bzlib.h>
+#include "CRC16Folding.h"